Remove legacy config classes support from extension annotation processor
This is a follow-up of https://github.com/quarkusio/quarkus/pull/48717
and should only be merged when 3.25 is branched.
