+++ core/processor/src/main/java/io/quarkus/annotation/processor/documentation/config/scanner/JavadocConfigMappingListener.java
@@ -1,8 +1,13 @@
 package io.quarkus.annotation.processor.documentation.config.scanner;
 
+import java.util.Optional;
+
+import javax.lang.model.element.Element;
+import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.TypeElement;
 
+import io.quarkus.annotation.processor.documentation.config.discovery.DiscoveryConfigRoot;
 import io.quarkus.annotation.processor.documentation.config.discovery.DiscoveryRootElement;
 import io.quarkus.annotation.processor.documentation.config.discovery.ParsedJavadoc;
 import io.quarkus.annotation.processor.documentation.config.discovery.ParsedJavadocSection;
@@ -17,19 +22,48 @@ import io.quarkus.annotation.processor.util.Utils;
 /**
  * This class is responsible for collecting and writing the Javadoc.
  */
-public class JavadocConfigMappingListener extends AbstractJavadocConfigListener {
+public class JavadocConfigMappingListener implements ConfigAnnotationListener {
+
+    private final Config config;
+    private final Utils utils;
+    private final ConfigCollector configCollector;
 
     JavadocConfigMappingListener(Config config, Utils utils, ConfigCollector configCollector) {
-        super(config, utils, configCollector);
+        this.config = config;
+        this.utils = utils;
+        this.configCollector = configCollector;
+    }
+
+    @Override
+    public Optional<DiscoveryConfigRoot> onConfigRoot(TypeElement configRoot) {
+        // we only get Javadoc for local classes
+        // classes coming from other modules won't have Javadoc available
+        if (!utils.element().isLocalClass(configRoot)) {
+            return Optional.empty();
+        }
+
+        Optional<String> rawJavadoc = utils.element().getJavadoc(configRoot);
+
+        if (rawJavadoc.isEmpty()) {
+            return Optional.empty();
+        }
+
+        ParsedJavadocSection parsedJavadocSection = JavadocUtil.parseConfigSectionJavadoc(rawJavadoc.get());
+        if (parsedJavadocSection.title() == null) {
+            return Optional.empty();
+        }
+
+        configCollector.addJavadocElement(
+                configRoot.getQualifiedName().toString(),
+                new JavadocElement(parsedJavadocSection.title(), parsedJavadocSection.format(), null,
+                        parsedJavadocSection.deprecated()));
+
+        return Optional.empty();
     }
 
     @Override
     public void onEnclosedMethod(DiscoveryRootElement discoveryRootElement, TypeElement clazz, ExecutableElement method,
             ResolvedType resolvedType) {
-        if (config.getExtension().isMixedModule() && !discoveryRootElement.isConfigMapping()) {
-            return;
-        }
-
         // we only get Javadoc for local classes
         // classes coming from other modules won't have Javadoc available
         if (!utils.element().isLocalClass(clazz)) {
@@ -64,6 +98,37 @@ public class JavadocConfigMappingListener extends AbstractJavadocConfigListener 
 
             configCollector.addJavadocElement(
                     clazz.getQualifiedName().toString() + Markers.DOT + method.getSimpleName().toString(),
+                    new JavadocElement(parsedJavadoc.description(), parsedJavadoc.format(), parsedJavadoc.since(),
+                            parsedJavadoc.deprecated()));
+        }
+    }
+
+    @Override
+    public void onResolvedEnum(TypeElement enumTypeElement) {
+        if (!utils.element().isLocalClass(enumTypeElement)) {
+            return;
+        }
+
+        for (Element enumElement : enumTypeElement.getEnclosedElements()) {
+            if (enumElement.getKind() != ElementKind.ENUM_CONSTANT) {
+                continue;
+            }
+
+            Optional<String> rawJavadoc = utils.element().getJavadoc(enumElement);
+
+            if (rawJavadoc.isEmpty()) {
+                continue;
+            }
+
+            ParsedJavadoc parsedJavadoc = JavadocUtil.parseConfigItemJavadoc(rawJavadoc.get());
+
+            if (parsedJavadoc.description() == null) {
+                continue;
+            }
+
+            configCollector.addJavadocElement(
+                    enumTypeElement.getQualifiedName().toString() + Markers.DOT + enumElement.getSimpleName()
+                            .toString(),
                     new JavadocElement(parsedJavadoc.description(), parsedJavadoc.format(), parsedJavadoc.since(),
                             parsedJavadoc.deprecated()));
         }
