+++ core/processor/src/main/java/io/quarkus/annotation/processor/util/Config.java
@@ -6,12 +6,10 @@ import io.quarkus.annotation.processor.documentation.config.model.ExtensionModul
 public class Config {
 
     private final ExtensionModule extensionModule;
-    private final boolean useConfigMapping;
     private final boolean debug;
 
-    public Config(ExtensionModule extensionModule, boolean useConfigMapping, boolean debug) {
+    public Config(ExtensionModule extensionModule, boolean debug) {
         this.extensionModule = extensionModule;
-        this.useConfigMapping = useConfigMapping;
         this.debug = debug;
     }
 
@@ -21,10 +19,6 @@ public class Config {
 
     public Extension getExtension() {
         return extensionModule.extension();
-    }
-
-    public boolean useConfigMapping() {
-        return useConfigMapping;
     }
 
     public boolean isDebug() {
